Ticket #374: use uintptr_t instead of uint32_t for auxParam of event dispatcher

+++ csrc/configuration/sopc_toolkit_config.c
@@ -71,7 +71,7 @@ static struct
 static SOPC_ComEvent_Fct* appFct = NULL;
 static SOPC_AddressSpaceNotif_Fct* pAddSpaceFct = NULL;
 
-void SOPC_Internal_ApplicationEventDispatcher(int32_t eventAndType, uint32_t id, void* params, uint32_t auxParam)
+void SOPC_Internal_ApplicationEventDispatcher(int32_t eventAndType, uint32_t id, void* params, uintptr_t auxParam)
 {
     switch (SOPC_AppEvent_AppEventType_Get(eventAndType))
     {
