+++ csrc/secure_channels/sopc_secure_channels_api_internal.h
@@ -26,7 +26,7 @@
 void SOPC_SecureChannels_EnqueueInternalEvent(SOPC_SecureChannels_InputEvent scEvent,
                                               uint32_t id,
                                               void* params,
-                                              uint32_t auxParam);
+                                              uintptr_t auxParam);
 
 /* Secure channel internal event enqueue function: event will be enqueued as next to be treated (only for close SC
  * situation) Note: it is important to close the SC as soon as possible in order to avoid any treatment of new messages
@@ -34,6 +34,6 @@ void SOPC_SecureChannels_EnqueueInternalEvent(SOPC_SecureChannels_InputEvent scE
 void SOPC_SecureChannels_EnqueueInternalEventAsNext(SOPC_SecureChannels_InputEvent scEvent,
                                                     uint32_t id,
                                                     void* params,
-                                                    uint32_t auxParam);
+                                                    uintptr_t auxParam);
 
 #endif /* SOPC_SECURE_CHANNELS_API_INTERNAL_H_ */
