+++ csrc/sockets/sopc_sockets_api.h
@@ -56,7 +56,7 @@ typedef enum {
 } SOPC_Sockets_InputEvent;
 
 /* Sockets event enqueue function */
-void SOPC_Sockets_EnqueueEvent(SOPC_Sockets_InputEvent socketEvent, uint32_t id, void* params, int32_t auxParam);
+void SOPC_Sockets_EnqueueEvent(SOPC_Sockets_InputEvent socketEvent, uint32_t id, void* params, uintptr_t auxParam);
 
 void SOPC_Sockets_Initialize(void);
