+++ tests/secure_channels/stubs/stub_sc_sopc_services_api.h
@@ -30,7 +30,7 @@ typedef struct SOPC_StubSC_ServicesEventParams
     SOPC_Services_Event event;
     uint32_t eltId;
     void* params;
-    uint32_t auxParam;
+    uintptr_t auxParam;
 } SOPC_StubSC_ServicesEventParams;
 
 // Async queue simulating the services dispatcher and containing SOPC_StubSC_ServicesEventParams elements
